Solve cosine-similarity

+++ cosine-similarity/cosine-similarity.py
@@ -15,4 +15,4 @@ def cosine_similarity(a, b):
     if norm_a == 0 or norm_b == 0:
         return 0.0
 
-    return float(np.dot(a, b) / (norm_a * norm_b))
+    return np.dot(a, b) / (norm_a * norm_b)
